requirements.txt
flask
transformers
torch
diffusers
fpdf
Pillow
nltk
